+++ Automation/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.1.2 (pre-release):
 * [Feature] Highlight the conditions that are being evaluated to "true" with green color.
+* [Feature] New scriptable components `Workplace`. Now, the number of workers in the workshop can be changed from the script.
 * [Change] Don't select the text when switching to the script input field.
 * [Fix] Chained building tool doesn't show the blocked completion icon.